Rename cron package
