+++ python_modules/libraries/dagster-cron/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dagster-cron-native"
+name = "dagster-cron"
 version = "0.1.0"
 edition = "2021"
 publish = false
