+++ python_modules/libraries/dagster-cron/tests/test_dagster_compatibility.py
@@ -5,7 +5,7 @@ import zoneinfo
 
 import pytest
 
-from dagster_cron_native import cron_string_iterator, is_valid_cron_string
+from dagster_cron import cron_string_iterator, is_valid_cron_string
 
 
 UTC = dt.timezone.utc
