+++ scripts/install_dev_python_modules.py
@@ -57,7 +57,7 @@ def main(
         "python_modules/dagster-test",
         "python_modules/dagster-webserver",
         "python_modules/dagit",
-        "python_modules/libraries/dagster-cron-native",
+        "python_modules/libraries/dagster-cron",
         "python_modules/libraries/dagster-shared",
         "python_modules/libraries/create-dagster",
         "python_modules/libraries/dagster-managed-elements",
