+++ ty/master/pyproject.toml
@@ -50,7 +50,7 @@ dependencies = [
   "dagster-clickhouse-pandas",
   "dagster-clickhouse-polars",
   "dagster-cloud-cli",
-  "dagster-cron-native",
+  "dagster-cron",
   "dagster-dask[yarn,pbs,kube]",
   "dagster-databricks",
   "dagster-datadog",
@@ -237,7 +237,7 @@ dagster-clickhouse          = { path = "../../python_modules/libraries/dagster-c
 dagster-clickhouse-pandas   = { path = "../../python_modules/libraries/dagster-clickhouse-pandas",   editable = true }
 dagster-clickhouse-polars   = { path = "../../python_modules/libraries/dagster-clickhouse-polars",   editable = true }
 dagster-cloud-cli           = { path = "../../python_modules/libraries/dagster-cloud-cli",           editable = true }
-dagster-cron-native         = { path = "../../python_modules/libraries/dagster-cron-native",         editable = true }
+dagster-cron         = { path = "../../python_modules/libraries/dagster-cron",         editable = true }
 dagster-dask                = { path = "../../python_modules/libraries/dagster-dask",                editable = true }
 dagster-databricks          = { path = "../../python_modules/libraries/dagster-databricks",          editable = true }
 dagster-datadog             = { path = "../../python_modules/libraries/dagster-datadog",             editable = true }
